pyproject.toml
[project]
name = "bug-tracker"
version = "0.1.0"
description = "Bug tracking tool with status lifecycle and automated tests"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.0",
    "pytest-playwright>=0.4.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
